removing pickUfePathPrim from ProxyShapeUI
This attempted to make ufe+USD selection work similar to maya, where
instead of selecting a shape, it selects the Transform above the shape.

However, trying to select the Xform above a Mesh in USD doesn't really
work the same, because in USD, a Mesh can have a Xform - and the
default export settings from maya will even "merge" the mesh + xform,
so there may not be ANY Xforms to find, and you end up selecting the
USD root.

(Also, this already wasn't present in the ProxyDrawOverride::userSelect,
so this gives consistent behavior as well.)